refactor: Refactor bots logic
`services/bots.py` is used to select the token for a repo / owner that will
authenticate requests to the git providers.

With recent changes to multi-github apps the logic has become quite convoluted and a bit
confusing. In particular with complex return structures of functions and an approach that
doesn't really take into consideration the owner's service.

These changes aim to separate the logic a bit better in different files grouped by "token type",
improve typehinting and docstrings, and also simplify the structure of data returned by some functions.
