+++ services/github.py
@@ -12,14 +12,14 @@ log = logging.getLogger(__name__)
 
 @cache.cache_function(ttl=480)
 def get_github_integration_token(
-    service,
-    integration_id=None,
+    service: str,
+    installation_id: int = None,
     app_id: Optional[str] = None,
     pem_path: Optional[str] = None,
 ):
     try:
         return _get_github_integration_token(
-            service, integration_id=integration_id, app_id=app_id, pem_path=pem_path
+            service, integration_id=installation_id, app_id=app_id, pem_path=pem_path
         )
     except InvalidInstallationError:
         log.warning("Failed to get installation token")
