+++ services/owner.py
@@ -2,6 +2,7 @@ import logging
 
 import shared.torngit as torngit
 from shared.config import get_config, get_verify_ssl
+from shared.django_apps.codecov_auth.models import Service
 from shared.typings.torngit import (
     GithubInstallationInfo,
     OwnerInfo,
@@ -9,7 +10,7 @@ from shared.typings.torngit import (
 )
 
 from helpers.token_refresh import get_token_refresh_callback
-from services.bots import get_owner_appropriate_bot_token, get_owner_installation_id
+from services.bots import get_github_app_info_for_owner, get_owner_appropriate_bot_token
 
 log = logging.getLogger(__name__)
 
@@ -21,11 +22,18 @@ def get_owner_provider_service(
         get_config("setup", "http", "timeouts", "connect", default=15),
         get_config("setup", "http", "timeouts", "receive", default=30),
     ]
-    service = owner.service
-    installation_info = get_owner_installation_id(
-        owner, using_integration, ignore_installation=ignore_installation
-    )
-    token = get_owner_appropriate_bot_token(owner, installation_dict=installation_info)
+    service = Service(owner.service)
+    installation_info = None
+    fallback_installations = None
+    if (
+        service in [Service.GITHUB, Service.GITHUB_ENTERPRISE]
+    ) and not ignore_installation:
+        installations_available_info = get_github_app_info_for_owner(
+            owner,
+        )
+        if installations_available_info != []:
+            installation_info, *fallback_installations = installations_available_info
+    token = get_owner_appropriate_bot_token(owner, installation_info)
     data = TorngitInstanceData(
         owner=OwnerInfo(
             service_id=owner.service_id, ownerid=owner.ownerid, username=owner.username
@@ -39,11 +47,11 @@ def get_owner_provider_service(
             app_id=installation_info.get("app_id"),
             pem_path=installation_info.get("pem_path"),
         )
-        data["fallback_installations"] = installation_info.get("fallback_installations")
+        data["fallback_installations"] = fallback_installations
 
     adapter_params = dict(
         token=token,
-        verify_ssl=get_verify_ssl(service),
+        verify_ssl=get_verify_ssl(service.value),
         timeouts=_timeouts,
         oauth_consumer_token=dict(
             key=get_config(service, "client_id"),
