+++ services/tests/test_repository_service.py
@@ -119,7 +119,7 @@ class TestRepositoryServiceTestCase(object):
         self, dbsession, mocker
     ):
         mocker.patch(
-            "services.bots.get_github_integration_token",
+            "services.bots.github_apps.get_github_integration_token",
             return_value="installation_token",
         )
         repo = RepositoryFactory.create(
